requirements.txt
selenium>=4.0.0
webdriver-manager>=3.8.0
pandas>=1.3.0
beautifulsoup4>=4.9.0
requests>=2.25.0
openai==1.93.0
python-dotenv==1.1.1
google-api-python-client>=2.0.0
gender-guesser 
langchain
tavily-python
flask>=2.0.0
flask-cors>=3.0.0 